requirements.txt
Pillow
python-mpd2
